reduce max size of fuzzed parameters
fixes the test failure noted in the previous commit (failure was due to 'overfiltering' of fuzzed inputs)

+++ src/test/Delegation.t.sol
@@ -105,7 +105,7 @@ contract DelegationTests is EigenLayerTestHelper {
         _testDelegation(operator, staker, ethAmount, eigenAmount, voteWeigher);
     }
 
-    function testUndelegation(address operator, address staker, uint256 ethAmount, uint256 eigenAmount)
+    function testUndelegation(address operator, address staker, uint96 ethAmount, uint96 eigenAmount)
         public
         fuzzedAddress(operator)
         fuzzedAddress(staker)
